+++ CHANGELOG.md
@@ -31,6 +31,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Astral characters are no longer encoded by dump/safeDump, #587.
 - "duplicate mapping key" exception now points at the correct column, #452.
+- Extra commas in flow collections (e.g. `[foo,,bar]`) now throw an exception
+  instead of producing null, #321.
 - Removed `bower.json`.